requirements.txt
pillow
joblib
requests
opencv-python
scikit-image
scikit-learn
tensorflow
matplotlib